refactor: reorganize utility modules and improve prompt function API
- Move cli.ts and prompt.ts to utils/ directory for better organization
- Move getCurrentUserInfo from utils/mod.ts to utils/prompt.ts
- Update getSystemPrompt to accept workingDirectory directly with optional UserInfo override
- Update module exports to reflect new file locations
