+++ src/mod.ts
@@ -3,10 +3,8 @@
 // Core agent
 export * from "./ZypherAgent.ts";
 export * from "./CheckpointManager.ts";
-export * from "./cli.ts";
 export * from "./error.ts";
 export * from "./message.ts";
-export * from "./prompt.ts";
 
 // Modules
 export * from "./llm/mod.ts";
