+++ src/utils/cli.ts
@@ -1,9 +1,9 @@
-import type { ZypherAgent } from "./ZypherAgent.ts";
+import type { ZypherAgent } from "../ZypherAgent.ts";
 import readline from "node:readline";
 import { stdin, stdout } from "node:process";
 import chalk from "chalk";
-import { printMessage } from "./message.ts";
-import { formatError } from "./error.ts";
+import { printMessage } from "../message.ts";
+import { formatError } from "../error.ts";
 import { eachValueFrom } from "rxjs-for-await";
 
 /**
